fix file layout - redirect unauthorized users to landing screen
