+++ app/(auth)/onboarding/index.tsx
