+++ app/(auth)/public/index.tsx
